Move core and wraper out of the way.
OperationBuilder has been clearly marked as an experiment in fluent api definition with examples from “wrapper” preserved in the README.md file.
