+++ modules/builder/src/main/java/org/eclipse/imagen/builder/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * A fluent fluent builder API for the creation of RenderedImages.
+ *
+ * This package is experimental and provided for early feedback and is not expected to be included in the initial release.
+ */
+package org.eclipse.imagen.builder;
