--- modules/core/src/main/java/org/eclipse/imagen/package-info.java
@@ -1,16 +1,0 @@
-/**
- * The Eclipse ImageN project provides an extensible, on-demand image processing library with no artificial
- * restrictions on raster size or number of bands.
- * <p>
- * ImageN provides:</p>
- * <ul>
- *     <li>Modern Java API using literate programming style</li>
- *     <li>Pure Java implementation</li>
- *     <li>Clear image processing operations, allowing installations to use native libs to accelerate processing if available</li>
- *     <li>On demand processing of large raster content staging tiles in memory for parallel processing</li>
- *     <li>No artificial limitation on raster size or number of bands to support multi-spectral imagery</li>
- * </ul>
- *
- * @since 1.0
- */
-package org.eclipse.imagen;
