+++ modules/pom.xml
@@ -12,8 +12,7 @@
     <packaging>pom</packaging>
     
     <modules>
-        <module>core</module>
-        <module>wrapper</module>
+        <module>builder</module>
     </modules>
     
     <build>
