--- modules/wrapper/src/main/java/org/eclipse/imagen/package-info.java
@@ -1,1 +1,0 @@
-package org.eclipse.imagen;
